Refactor Formulation to a separate module
