+++ cooper/state_logger.py
@@ -2,7 +2,7 @@ from collections import OrderedDict
 from copy import deepcopy
 from typing import List
 
-from .problem import Formulation
+from .formulation import Formulation
 
 
 class StateLogger:
